move benchmark code in `test/editorPieceTable.test.ts` to `scripts/benchmarkEditorPieceTable.ts`
